Change step.sequence_number from float to int
The idea of float was to allow inserts without updating the other
numbers. However, having sequence number be the integer sequence
position is nice for e.g. URLs and straightforward view code.

+++ backend/app/models/__init__.py
@@ -93,7 +93,7 @@ class ProcessStep(BaseModel):
         Process, on_delete=models.deletion.PROTECT, related_name="steps"
     )
     service = models.ForeignKey(Service, on_delete=models.deletion.CASCADE)
-    sequence_number = models.FloatField()
+    sequence_number = models.PositiveIntegerField()
 
 
 class Provider(BaseModel):
